+++ src/common/dwt.c
@@ -151,7 +151,7 @@ static void dwt_decompose_vert(float *const restrict out, const float *const res
     // i.e. we move as many rows in from the edge as we would have been beyond the edge
     // for the top edge, this means we can simply use the absolute value of row-vscale; for the bottom edge,
     //   we need to reflect around height
-    const size_t rowstart = 4 * row * width;
+    const size_t rowstart = (size_t)4 * row * width;
     const int below_row = (row + vscale < height) ? (row + vscale) : 2*(height-1) - (row + vscale);
     const float* const restrict center = in + rowstart;
     const float* const restrict above = in + 4 * abs(row - vscale) * width;
@@ -280,7 +280,7 @@ static void dwt_wavelet_decompose(float *img, dwt_params_t *const p, _dwt_layer_
 
   if(p->merge_from_scale > 0)
   {
-    merged_layers = dt_alloc_align_float(p->width * p->height * p->ch);
+    merged_layers = dt_alloc_align_float((size_t)p->width * p->height * p->ch);
     if(merged_layers == NULL)
     {
       printf("not enough memory for wavelet decomposition");
@@ -429,7 +429,7 @@ static void dwt_denoise_vert_1ch(float *const restrict out, const float *const r
     // i.e. we move as many rows in from the edge as we would have been beyond the edge
     // for the top edge, this means we can simply use the absolute value of row-vscale; for the bottom edge,
     //   we need to reflect around height
-    const size_t rowstart = row * width;
+    const size_t rowstart = (size_t)row * width;
     const int below_row = (row + vscale < height) ? (row + vscale) : 2*(height-1) - (row + vscale);
     const float *const restrict center = in + rowstart;
     const float *const restrict above =  in + abs(row - vscale) * width;
@@ -534,7 +534,7 @@ static void dwt_denoise_horiz_1ch(float *const restrict out, float *const restri
  */
 void dwt_denoise(float *const img, const int width, const int height, const int bands, const float *const noise)
 {
-  float *const details = dt_alloc_align_float(2 * width * height);
+  float *const details = dt_alloc_align_float((size_t)2 * width * height);
   float *const interm = details + width * height;	// temporary storage for use during each pass
 
   // zero the accumulator
