+++ src/iop/exposure.c
@@ -456,7 +456,7 @@ void process(struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, const 
   float *const restrict out = (float*)o;
   const float black = d->black;
   const float scale = d->scale;
-  const size_t npixels = roi_out->width * roi_out->height;
+  const size_t npixels = (size_t)roi_out->width * roi_out->height;
 #ifdef _OPENMP
 #pragma omp parallel for simd default(none) \
   dt_omp_firstprivate(ch, npixels, black, scale, in, out)  \
